+++ lib/inquisitio/searcher.rb
@@ -26,17 +26,24 @@ module Inquisitio
       results
     end
 
-=begin
     def ids
-      @ids ||= to_a.map{|result|result['id']}
+      @ids ||= map{|r|r['med_id']}
     end
 
     def records
-      @records ||= to_a.map do |result|
-        {result['type'] => result['id']}
+      @records ||= begin
+        klasses = {}
+        map do |result|
+          klass = result['med_type']
+          klasses[klass] ||= []
+          klasses[klass] << result['med_id']
+        end
+
+        klasses.map {|klass, ids|
+          constantize(klass).where(id: ids)
+        }.flatten
       end
     end
-=end
 
     def where(value)
       clone do |s|
@@ -103,20 +110,50 @@ module Inquisitio
     end
 
     def search_url
-      @search_url ||= SearchUrlBuilder.build(
-        query: params[:criteria],
-        filters: params[:filters],
-        arguments: params[:with].merge({
-          size: params[:per],
-          start: params[:per] * params[:page]
-        }),
-        return_fields: params[:returns]
-      )
+      @search_url ||= begin
+        return_fields = params[:returns].empty?? [:med_type, :med_id] : params[:returns]
+
+        SearchUrlBuilder.build(
+          query: params[:criteria],
+          filters: params[:filters],
+          arguments: params[:with].merge({
+            size: params[:per],
+            start: params[:per] * params[:page]
+          }),
+          return_fields: return_fields
+        )
+      end
     end
 
     def clone
       Searcher.new(DeepClone.clone(params)) do |s|
         yield(s) if block_given?
+      end
+    end
+
+    def constantize(camel_cased_word)
+      names = camel_cased_word.gsub("_", "::").split('::')
+      names.shift if names.empty? || names.first.empty?
+
+      names.inject(Object) do |constant, name|
+        if constant == Object
+          constant.const_get(name)
+        else
+          candidate = constant.const_get(name)
+          next candidate if constant.const_defined?(name, false)
+          next candidate unless Object.const_defined?(name)
+
+          # Go down the ancestors to check it it's owned
+          # directly before we reach Object or the end of ancestors.
+          constant = constant.ancestors.inject do |const, ancestor|
+            break const    if ancestor == Object
+            break ancestor if ancestor.const_defined?(name, false)
+            const
+          end
+
+          # owner is in Object, so raise
+          constant.const_get(name, false)
+        end
       end
     end
   end