+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 0.0.13 / Unreleased
 
+[FEATURE] Implement Searcher#records
+[FEATURE] Implement Searcher#ids
 [FEATURE] Allow for Array-style method iteration
 [FEATURE] Changed syntax to match ActiveRecord style.